feat: add ProjectCaseStudyContent and ProjectHeader components; refactor Projects and ProjectPage to utilize new components
